@@ -23,9 +23,9 @@
   },
   "homepage": "https://github.com/ldc4/ldc4-pages#readme",
   "dependencies": {
-    "next": "^7.0.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-spring": "^7.2.10"
+    "next": "^8.1.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-spring": "^8.0.20"
   }
 }
